+++ tests/unit/vector/runVectorHandlerTests.cpp
@@ -25,7 +25,7 @@ int main(int, char**)
     result += test.axpyMulti(100, 10);
     result += test.massDot(100, 10);
     result += test.scale(100);
-    result += test.scaleInv(100);
+    result += test.diagSolve(100);
     result += test.max(100);
     result += test.abs(100);
 
@@ -51,7 +51,7 @@ int main(int, char**)
     result += test.massDot(1000, 30);
     result += test.amax(1000);
     result += test.scale(1000);
-    result += test.scaleInv(1000);
+    result += test.diagSolve(1000);
     result += test.max(1000);
     result += test.abs(1000);
 
@@ -78,7 +78,7 @@ int main(int, char**)
     result += test.massDot(1000, 30);
     result += test.amax(1000);
     result += test.scale(1000);
-    result += test.scaleInv(1000);
+    result += test.diagSolve(1000);
     result += test.max(1000);
     result += test.abs(1000);